requirements.txt
# LangChain Core
langchain
langchain-core
langchain-community

# OpenAI Integration
langchain-openai
openai



# Google Gemini (PaLM) Integration
langchain-google-genai
google-generativeai



# Environment Variable Management
python-dotenv

# Machine Learning Utilities
numpy
scikit-learn
streamlit